requirements.txt
python-dotenv
flask
flask_socketio
flask_bootstrap
flask-mqtt
eventlet